Separated out unit tests into different classes and made test driver.
